+++ test/unit/fuzz-tools.test.js
@@ -442,6 +442,7 @@ import * as waldenstromPerthesV463 from '../../lib/waldenstrom-perthes-v463.js';
 import * as crawfordTaaaV464 from '../../lib/crawford-taaa-v464.js';
 import * as stameyIncontinenceV465 from '../../lib/stamey-incontinence-v465.js';
 import * as letournelAcetabulumV466 from '../../lib/letournel-acetabulum-v466.js';
+import * as bromageScaleV467 from '../../lib/bromage-scale-v467.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -829,6 +830,7 @@ const MODULES = {
   'crawford-taaa-v464.js': crawfordTaaaV464,
   'stamey-incontinence-v465.js': stameyIncontinenceV465,
   'letournel-acetabulum-v466.js': letournelAcetabulumV466,
+  'bromage-scale-v467.js': bromageScaleV467,
   'tb-testing.js': tbTesting,
 };